Define bytecode chunks as having a maximum size of u32::MAX
It's good to be precise and avoid using `usize` for things that should really be
platform independent, and this also has the benefit of reducing the size of
SimpleFunctionInfo to 16 bytes.

+++ core/bytecode/src/instruction_reader.rs
@@ -173,7 +173,7 @@ pub enum Instruction {
     SimpleFunction {
         register: u8,
         arg_count: u8,
-        size: usize,
+        size: u16,
     },
     Function {
         register: u8,
@@ -182,7 +182,7 @@ pub enum Instruction {
         variadic: bool,
         generator: bool,
         arg_is_unpacked_tuple: bool,
-        size: usize,
+        size: u16,
     },
     Capture {
         function: u8,
@@ -273,18 +273,18 @@ pub enum Instruction {
         rhs: u8,
     },
     Jump {
-        offset: usize,
+        offset: u16,
     },
     JumpBack {
-        offset: usize,
+        offset: u16,
     },
     JumpIfTrue {
         register: u8,
-        offset: usize,
+        offset: u16,
     },
     JumpIfFalse {
         register: u8,
-        offset: usize,
+        offset: u16,
     },
     Call {
         result: u8,
@@ -315,7 +315,7 @@ pub enum Instruction {
     IterNext {
         result: Option<u8>,
         iterator: u8,
-        jump_offset: usize,
+        jump_offset: u16,
         temporary_output: bool,
     },
     TempIndex {
@@ -388,7 +388,7 @@ pub enum Instruction {
     },
     TryStart {
         arg_register: u8,
-        catch_offset: usize,
+        catch_offset: u16,
     },
     TryEnd,
     Debug {
@@ -1066,7 +1066,7 @@ impl Iterator for InstructionReader {
             Op::SimpleFunction => {
                 let register = get_u8!();
                 let arg_count = get_u8!();
-                let size = get_u16!() as usize;
+                let size = get_u16!();
 
                 Some(SimpleFunction {
                     register,
@@ -1079,7 +1079,7 @@ impl Iterator for InstructionReader {
                 let arg_count = get_u8!();
                 let capture_count = get_u8!();
                 let flags = FunctionFlags::from_byte(get_u8!());
-                let size = get_u16!() as usize;
+                let size = get_u16!();
 
                 Some(Function {
                     register,
@@ -1179,19 +1179,15 @@ impl Iterator for InstructionReader {
                 lhs: get_u8!(),
                 rhs: get_u8!(),
             }),
-            Op::Jump => Some(Jump {
-                offset: get_u16!() as usize,
-            }),
-            Op::JumpBack => Some(JumpBack {
-                offset: get_u16!() as usize,
-            }),
+            Op::Jump => Some(Jump { offset: get_u16!() }),
+            Op::JumpBack => Some(JumpBack { offset: get_u16!() }),
             Op::JumpIfTrue => Some(JumpIfTrue {
                 register: get_u8!(),
-                offset: get_u16!() as usize,
+                offset: get_u16!(),
             }),
             Op::JumpIfFalse => Some(JumpIfFalse {
                 register: get_u8!(),
-                offset: get_u16!() as usize,
+                offset: get_u16!(),
             }),
             Op::Call => Some(Call {
                 result: get_u8!(),
@@ -1222,19 +1218,19 @@ impl Iterator for InstructionReader {
             Op::IterNext => Some(IterNext {
                 result: Some(get_u8!()),
                 iterator: get_u8!(),
-                jump_offset: get_u16!() as usize,
+                jump_offset: get_u16!(),
                 temporary_output: false,
             }),
             Op::IterNextTemp => Some(IterNext {
                 result: Some(get_u8!()),
                 iterator: get_u8!(),
-                jump_offset: get_u16!() as usize,
+                jump_offset: get_u16!(),
                 temporary_output: true,
             }),
             Op::IterNextQuiet => Some(IterNext {
                 result: None,
                 iterator: get_u8!(),
-                jump_offset: get_u16!() as usize,
+                jump_offset: get_u16!(),
                 temporary_output: false,
             }),
             Op::IterUnpack => Some(IterNext {
@@ -1368,7 +1364,7 @@ impl Iterator for InstructionReader {
             }),
             Op::TryStart => Some(TryStart {
                 arg_register: get_u8!(),
-                catch_offset: get_u16!() as usize,
+                catch_offset: get_u16!(),
             }),
             Op::TryEnd => Some(TryEnd),
             Op::Debug => Some(Debug {
